Changed .html to .php and centralized navigation bar
Changed everything to use .php. This allows for common elements, like
the navigation bar, to exist in a single file but included in other .php
files. If edits need to be made, only that single file will need to be
changed (as opposed to every .html file with the element).